# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Gin-based API for the Fox \u0026 Hound application.",
        "title": "Fox \u0026 Hound API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/head-data": {
            "get": {
                "description": "Fetches head data for certain candidate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "head-data"
                ],
                "summary": "Get head data for candidate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Email",
                        "name": "userEmail",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HeadDataDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Stores or updates head data in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "head-data"
                ],
                "summary": "Save head data",
                "parameters": [
                    {
                        "description": "head data",
                        "name": "headData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.HeadDataDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message: HeadData saved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to save headData",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/management/agreed-candidates": {
            "get": {
                "description": "Fetches head data for agreed candidates",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "head-data"
                ],
                "summary": "Get agreed candidates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.HeadDataDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/management/average": {
            "get": {
                "description": "Fetches management average for cards",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "management-average"
                ],
                "summary": "Get management average for cards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ManagementAverageDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/management/summary": {
            "get": {
                "description": "Fetches management summary for certain candidate",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "management-summary"
                ],
                "summary": "Get management summary for candidate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Email",
                        "name": "userEmail",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ManagementSummaryDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/rating-cards": {
            "get": {
                "description": "Retrieves all rating cards from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rating-card"
                ],
                "summary": "Get all rating cards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.RatingCard"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ratings/average": {
            "get": {
                "description": "Fetches average for every card - agreed on flag is a must",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "head-data"
                ],
                "summary": "Get average per card",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AverageRatingDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ratings/candidate": {
            "get": {
                "description": "Fetches rating cards and enriches them with existing ratings for a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rating-candidate"
                ],
                "summary": "Get candidate ratings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Email",
                        "name": "userEmail",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CandidateRatingDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Stores or updates candidate ratings in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rating-candidate"
                ],
                "summary": "Save candidate ratings",
                "parameters": [
                    {
                        "description": "List of candidate ratings",
                        "name": "ratings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CandidateRatingDTO"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message: Ratings saved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to save ratings",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ratings/employer": {
            "get": {
                "description": "Fetches rating cards and enriches them with existing ratings for a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rating-employer"
                ],
                "summary": "Get candidate ratings for employer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Email",
                        "name": "userEmail",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EmployerRatingDTO"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Stores or updates candidate ratings of employer in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rating-employer"
                ],
                "summary": "Save employer ratings",
                "parameters": [
                    {
                        "description": "List of candidate ratings",
                        "name": "ratings",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EmployerRatingDTO"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message: Ratings saved successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to save ratings",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/ratings/employer/candidates": {
            "get": {
                "description": "Fetches rating candidates",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rating-employer"
                ],
                "summary": "Get rating candidates",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AverageRatingDTO": {
            "type": "object",
            "properties": {
                "NumberOfAgreedRatings": {
                    "type": "integer"
                },
                "average": {
                    "type": "number"
                },
                "ratingCardId": {
                    "type": "integer"
                }
            }
        },
        "models.CandidateRatingDTO": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "notApplicableCandidate": {
                    "type": "boolean"
                },
                "orderId": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "ratingCandidate": {
                    "type": "integer"
                },
                "ratingCardId": {
                    "type": "integer"
                },
                "textResponseCandidate": {
                    "type": "string"
                },
                "timeStampCandidate": {
                    "type": "string"
                },
                "userEmail": {
                    "type": "string"
                }
            }
        },
        "models.CategoryEnum": {
            "type": "string",
            "enum": [
                "Performance",
                "Technical Skillset",
                "Technical Predispositions",
                "Sales",
                "Recruiting",
                "Teamwork",
                "Coaching",
                "Prodyna Insights",
                "Overall"
            ],
            "x-enum-varnames": [
                "CategoryPerformance",
                "CategoryTechnicalSkillset",
                "CategoryTechnicalPredisposition",
                "CategorySales",
                "CategoryRecruiting",
                "CategoryTeamwork",
                "CategoryCoaching",
                "CategoryProdynaInsights",
                "CategoryOverall"
            ]
        },
        "models.EmployerRatingDTO": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "notApplicableCandidate": {
                    "type": "boolean"
                },
                "notApplicableEmployer": {
                    "type": "boolean"
                },
                "orderId": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "ratingCandidate": {
                    "type": "integer"
                },
                "ratingCardId": {
                    "type": "integer"
                },
                "ratingEmployer": {
                    "type": "integer"
                },
                "textResponseCandidate": {
                    "type": "string"
                },
                "textResponseEmployer": {
                    "type": "string"
                },
                "timeStampCandidate": {
                    "type": "string"
                },
                "timeStampEmployer": {
                    "type": "string"
                },
                "userEmail": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.HeadDataDTO": {
            "type": "object",
            "properties": {
                "abstract": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "agreedOn": {
                    "type": "boolean"
                },
                "experienceSince": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "startAtProdyna": {
                    "type": "string"
                },
                "userEmail": {
                    "type": "string"
                }
            }
        },
        "models.ManagementAverageDTO": {
            "type": "object",
            "properties": {
                "average": {
                    "type": "number"
                },
                "category": {
                    "$ref": "#/definitions/models.CategoryEnum"
                }
            }
        },
        "models.ManagementSummaryDTO": {
            "type": "object",
            "properties": {
                "ratings": {
                    "description": "Array of management ratings",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ManagementSummaryRatingDTO"
                    }
                },
                "userEmail": {
                    "type": "string"
                }
            }
        },
        "models.ManagementSummaryRatingDTO": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/models.CategoryEnum"
                },
                "rating": {
                    "type": "number"
                }
            }
        },
        "models.RatingCard": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "orderId": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                }
            }
        }
    }
}